+++ tests/common/paths.spec.ts
@@ -1,3 +1,4 @@
+import { normalize } from 'path';
 import { paths } from '../../src/common/paths';
 
 describe('paths', () => {
@@ -7,7 +8,7 @@ describe('paths', () => {
         );
 
         expect(folderPathRaw).toEqual('./example/my/');
-        expect(folderPathNormal).toEqual('example/my');
+        expect(folderPathNormal).toEqual(normalize('example/my/'));
         expect(specFileName).toEqual('my.component.spec.ts');
         expect(fileName).toEqual('my.component');
     });
@@ -18,7 +19,7 @@ describe('paths', () => {
         );
 
         expect(folderPathRaw).toEqual('.\\example\\my\\');
-        expect(folderPathNormal).toEqual('example/my');
+        expect(folderPathNormal).toEqual(normalize('example/my/'));
         expect(specFileName).toEqual('my.component.spec.ts');
         expect(fileName).toEqual('my.component');
     });
