+++ tests/spec/test-data/update-custom-mark/to-update.component.ts
@@ -1,0 +1,11 @@
+import { Router } from '@the/router';
+import { Just } from 'maybe';
+import { Service } from './service';
+export class ToUpdateComponent {
+  constructor(service: Service, router: Router, just: Just) {}
+
+  myMethod() {}
+
+  mySecondMethod() {}
+}
+
